+++ .vscode/typescript.code-snippets
@@ -4,13 +4,19 @@
     "description": "Generate a complete endpoint index.ts file",
     "body": [
       "import { ResourceClient } from '../../core/resource-client.ts';",
+      "import type {",
+      "  Create${1:Resource},",
+      "  Get${1:Resource}Params,",
+      "  Get${1:Resource}sParams,",
+      "  ${1:Resource},",
+      "  Update${1:Resource},",
+      "} from './types.ts';",
+      "import type { HttpResponse, PaginatedHttpResponse, PaginatedResponse } from 'types/http.ts';",
+      "import type { Options } from 'types/request-building-options.ts';",
       "import type { RequestHandler } from '../../core/request-handler.ts';",
-      "import type { PaginatedResponse } from '../../core/types/endpoint/response.ts';",
-      "import type { Options } from '../../core/types/internal/request-building-options.ts';",
-      "import type { Get${1:Resource}Params, Get${1:Resource}sParams, ${1:Resource} } from './types.ts';",
       "",
       "/**",
-      " * Provides access to the ${2:${1/(.*)/${1:/downcase}/}} endpoints of the xMatters API.",
+      " * Provides access to the ${2:${1/(.*)/${1:/downcase}/}}s endpoints of the xMatters API.",
       " * Use this class to manage ${2:${1/(.*)/${1:/downcase}/}}s, including listing, creating, updating, and deleting ${2:${1/(.*)/${1:/downcase}/}}s.",
       " */",
       "export class ${1:Resource}sEndpoint {",
@@ -28,7 +34,9 @@
       "   * @returns The HTTP response containing a paginated list of ${2:${1/(.*)/${1:/downcase}/}}s",
       "   * @throws {XmApiError} If the request fails",
       "   */",
-      "  get(options?: Options & { query?: Get${1:Resource}sParams }) {",
+      "  get(",
+      "    options?: Options & { query?: Get${1:Resource}sParams },",
+      "  ): Promise<PaginatedHttpResponse<${1:Resource}>> {",
       "    return this.http.get<PaginatedResponse<${1:Resource}>>(options);",
       "  }",
       "",
@@ -40,7 +48,10 @@
       "   * @returns The HTTP response containing the ${2:${1/(.*)/${1:/downcase}/}}",
       "   * @throws {XmApiError} If the request fails",
       "   */",
-      "  getByIdentifier(identifier: string, options?: Options & { query?: Get${1:Resource}Params }) {",
+      "  getByIdentifier(",
+      "    identifier: string,",
+      "    options?: Options & { query?: Get${1:Resource}Params },",
+      "  ): Promise<HttpResponse<${1:Resource}>> {",
       "    return this.http.get<${1:Resource}>({ ...options, path: identifier });",
       "  }",
       "",
@@ -52,7 +63,10 @@
       "   * @returns The HTTP response containing the created or updated ${2:${1/(.*)/${1:/downcase}/}}",
       "   * @throws {XmApiError} If the request fails",
       "   */",
-      "  save(${2:${1/(.*)/${1:/downcase}/}}: Partial<${1:Resource}>, options?: Options) {",
+      "  save(",
+      "    ${2:${1/(.*)/${1:/downcase}/}}: Create${1:Resource} | Update${1:Resource},",
+      "    options?: Options,",
+      "  ): Promise<HttpResponse<${1:Resource}>> {",
       "    return this.http.post<${1:Resource}>({ ...options, body: ${2:${1/(.*)/${1:/downcase}/}} });",
       "  }",
       "",
@@ -64,8 +78,8 @@
       "   * @returns The HTTP response",
       "   * @throws {XmApiError} If the request fails",
       "   */",
-      "  delete(id: string, options?: Options) {",
-      "    return this.http.delete<void>({ ...options, path: id });",
+      "  delete(id: string, options?: Options): Promise<HttpResponse<${1:Resource}>> {",
+      "    return this.http.delete<${1:Resource}>({ ...options, path: id });",
       "  }",
       "}",
       "$0"
