+++ deno.json
@@ -3,7 +3,8 @@
   "version": "0.0.1",
   "exports": "./src/index.ts",
   "imports": {
-    "std/": "https://deno.land/std@0.224.0/"
+    "std/": "https://deno.land/std@0.224.0/",
+    "types/": "./src/core/types/"
   },
   "tasks": {
     "cache": "DENO_TLS_CA_STORE=system deno cache --reload src/**/*.ts",
