+++ src/core/defaults/http-client.ts
@@ -1,4 +1,4 @@
-import type { Headers, HttpClient, HttpRequest, HttpResponse } from '../types/internal/http.ts';
+import type { Headers, HttpClient, HttpRequest, HttpResponse } from 'types/http.ts';
 
 export class DefaultHttpClient implements HttpClient {
   async send(request: HttpRequest): Promise<HttpResponse> {
