+++ src/core/types/config.ts
