+++ src/core/types/mutable-auth-state.ts
