+++ src/endpoints/groups/index.test.ts
@@ -1,6 +1,6 @@
 import { GroupsEndpoint } from './index.ts';
-import { RequestHandler } from '../../core/request-handler.ts';
 import { MockHttpClient, MockLogger, TestConstants } from '../../core/test-utils.ts';
+import { RequestHandler } from '../../core/request-handler.ts';
 
 // Shared test infrastructure - MockHttpClient auto-resets between tests
 const mockHttpClient = new MockHttpClient();
