+++ src/endpoints/oauth/index.ts
@@ -1,8 +1,8 @@
-import type { RequestHandler } from '../../core/request-handler.ts';
-import type { HttpResponse } from '../../core/types/internal/http.ts';
-import type { OAuth2TokenResponse } from '../../core/types/internal/oauth.ts';
+import { AuthType } from 'types/mutable-auth-state.ts';
 import { XmApiError } from '../../core/errors.ts';
-import { AuthType } from '../../core/types/internal/mutable-auth-state.ts';
+import type { HttpResponse } from 'types/http.ts';
+import type { OAuth2TokenResponse } from 'types/oauth.ts';
+import type { RequestHandler } from '../../core/request-handler.ts';
 
 export class OAuthEndpoint {
   constructor(
