+++ src/index.test.ts
@@ -1,6 +1,6 @@
 import { expect } from 'std/expect/mod.ts';
-import { XmApi, XmApiError } from './index.ts';
 import { MockHttpClient, MockLogger, withFakeTime } from './core/test-utils.ts';
+import { XmApi, XmApiError } from './index.ts';
 
 const mockHttpClient = new MockHttpClient();
 const mockLogger = new MockLogger();
